Met Example of p-value comparison
Case Study: reanalysis daily-mean wind speed in Honolulu (JJA 1960). Does this follow normal or weibull distribution?
K-S => neither can be rejected in this case.
Idea of this case study was to examine differences in p-value and to try to understand if there was any point at which differences may be meaningful. For instance, if for 1000 runs of a subsampled dataset, one distribution always has a higher p-value than another is this meaningful? This study was the basis for a question I uploaded to Stack Exchange / Cross Validated.