Update LLR p-value; L1 + L2 bug fixes
The Vuong Log-Likelihood Ratio test evaluates which of two distributions fits a data set better. It gives a result R which if positive favours one distribution, if negative the other. The further from zero R is, the less likely the result is due to random fluctuations. To quantify this somewhat we apply a p-value. When p < alpha, the result is significant and we can trust the result for R.
Clauset et al (2009) recommend a "conservative" alpha of 0.1 instead of 0.05. So if p < 0.1 , the result is significant. This removes a bit of the noise...